lmao code worky!!!1!!
so my flexbox code wasn't working. wanna know why?
i was inserting the cards directly into the body (it was temporary but i don't remember why i didn't implement it.). the flexbox css was for the div that was intended to have the cards. the code straight up wasn't associated with the correct thing. annoying.